requirements.txt
python-dotenv
pandas
numpy 
sqlalchemy
certifi
psycopg2-binary
scikit-learn
pyaml

-e .